pyproject.toml
[project]
name = "mtg_deck_editor"
version = "0.0.1"
description = "A practice flask application."
readme = "README.md"
license = {file = "LICENSE"}
dependencies = [
    "Flask>=3.1.0",
    "Flask-SQLAlchemy>=3.1.1"
]
requires-python = ">=3.13.0"